refactor: Move breadcrumb feature into its own space
