+++ src/modules/breadcrumb/components/MobileAwareBreadcrumb.jsx
@@ -2,8 +2,8 @@ import React from 'react'
 
 import useBreakpoints from 'cozy-ui/transpiled/react/providers/Breakpoints'
 
-import DesktopBreadcrumb from './components/DesktopBreadcrumb/DesktopBreadcrumb'
-import MobileBreadcrumb from './components/MobileBreadcrumb/MobileBreadcrumb'
+import DesktopBreadcrumb from 'modules/breadcrumb/components/DesktopBreadcrumb'
+import MobileBreadcrumb from 'modules/breadcrumb/components/MobileBreadcrumb'
 
 export const MobileAwareBreadcrumb = props => {
   const { isMobile } = useBreakpoints()
