+++ src/modules/breadcrumb/components/MobileBreadcrumb.jsx
@@ -2,8 +2,8 @@ import React, { useCallback } from 'react'
 
 import { BarCenter, BarLeft } from 'cozy-bar'
 
-import Breadcrumb from '../../Breadcrumb'
 import BackButton from 'components/Button/BackButton'
+import Breadcrumb from 'modules/breadcrumb/components/Breadcrumb'
 
 const MobileBreadcrumb = ({ onBreadcrumbClick, path, ...props }) => {
   const navigateBack = useCallback(() => {
