+++ src/modules/views/Folder/FolderViewBreadcrumb.jsx
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types'
 import React, { useCallback } from 'react'
 
 import { useBreadcrumbPath } from './hooks/useBreadcrumbPath.jsx'
-import { MobileAwareBreadcrumb as Breadcrumb } from 'modules/navigation/Breadcrumb/MobileAwareBreadcrumb'
+import { MobileAwareBreadcrumb as Breadcrumb } from 'modules/breadcrumb/components/MobileAwareBreadcrumb'
 
 const FolderViewBreadcrumb = ({
   currentFolderId,
